+++ src/main/fc/rc_controls.h
@@ -109,7 +109,6 @@ bool checkStickPosition(stickPositions_e stickPos);
 bool areSticksInApModePosition(uint16_t ap_mode);
 bool areSticksDeflected(void);
 bool isRollPitchStickDeflected(uint8_t deadband);
-uint16_t setDesiredThrottle(uint16_t throttle, bool allowMotorStop);
 throttleStatus_e calculateThrottleStatus(throttleStatusType_e type);
 int16_t throttleStickMixedValue(void);
 rollPitchStatus_e calculateRollPitchCenterStatus(void);
